requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
Werkzeug==2.0.3
python-dotenv==0.21.0
pymongo==3.12.3
dnspython==2.3.0
oauthlib==3.2.2
openai==0.27.0
requests==2.26.0
textblob==0.17.1
transformers==4.30.2
torch==1.13.1
sentencepiece==0.1.99 
protobuf==3.20.3
gunicorn==20.1.0       
